bugfixes for cyclic / type-checking imports
