+++ ics/alarm/__init__.py
@@ -1,7 +1,8 @@
 from ics.alarm.audio import AudioAlarm
+from ics.alarm.base import BaseAlarm
 from ics.alarm.custom import CustomAlarm
 from ics.alarm.display import DisplayAlarm
 from ics.alarm.email import EmailAlarm
 from ics.alarm.none import NoneAlarm
 
-__all__ = ["AudioAlarm", "DisplayAlarm", "EmailAlarm", "NoneAlarm", "CustomAlarm"]
+__all__ = ["BaseAlarm", "AudioAlarm", "DisplayAlarm", "EmailAlarm", "NoneAlarm", "CustomAlarm"]
