+++ ics/alarm/utils.py
@@ -1,21 +1,21 @@
-from ics.alarm.audio import AudioAlarm
-from ics.alarm.custom import CustomAlarm
-from ics.alarm.display import DisplayAlarm
-from ics.alarm.email import EmailAlarm
-from ics.alarm.none import NoneAlarm
 from ics.utils import get_lines
 
 
 def get_type_from_action(action_type):
     if action_type == "DISPLAY":
+        from ics.alarm import DisplayAlarm
         return DisplayAlarm
     elif action_type == "AUDIO":
+        from ics.alarm import AudioAlarm
         return AudioAlarm
     elif action_type == "NONE":
+        from ics.alarm import NoneAlarm
         return NoneAlarm
     elif action_type == 'EMAIL':
+        from ics.alarm import EmailAlarm
         return EmailAlarm
     else:
+        from ics.alarm import CustomAlarm
         return CustomAlarm
