+++ ics/parsers/attendee_parser.py
@@ -1,5 +1,10 @@
+from typing import TYPE_CHECKING
+
 from ics.parsers.parser import Parser
 from ics.utils import unescape_string
+
+if TYPE_CHECKING:
+    from ics.attendee import Person, Attendee
 
 
 class PersonParser(Parser):
